ended up in circular imports, just made a new utils file instead
