+++ sktime_dl/deeplearning/tests/test_classifiers.py
@@ -1,7 +1,7 @@
 from sktime.datasets import load_italy_power_demand, load_basic_motions
 
 from sktime_dl.deeplearning import CNNClassifier
-from sktime_dl.utils import construct_all_classifiers, SMALL_NB_EPOCHS
+from sktime_dl.utils.model_lists import construct_all_classifiers, SMALL_NB_EPOCHS
 
 
 def test_basic_univariate(network=CNNClassifier(nb_epochs=SMALL_NB_EPOCHS)):
