+++ sktime_dl/utils/__init__.py
@@ -1,4 +1,2 @@
-from ._models import save_trained_model, check_is_fitted, construct_all_classifiers, construct_all_regressors
+from ._models import save_trained_model, check_is_fitted
 from ._data import check_and_clean_data
-
-SMALL_NB_EPOCHS = 3
